workflows/containers/pyproject.toml
[tool.poetry]
name = "tools-artifacts"
version = "0.1.0"
description = ""
authors = ["yuuki <yuki.tsubo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-storage = "^1.39"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
